pyproject.toml
[project]
name = "mcp-client-basics"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.40.0",
    "chromadb>=1.0.15",
    "fastmcp>=2.10.6",
    "google-genai>=1.27.0",
    "langchain-text-splitters>=0.3.9",
    "mcp>=1.1.1",
    "openai>=1.97.1",
    "python-dotenv>=1.0.1",
]
